requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.4.2
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
openai-whisper==20240930
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
pillow==10.4.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
qrcode==8.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
yt-dlp==2025.4.30
